+++ pkg/gardenlet/controller/seed/seed/reconciler_reconcile.go
@@ -238,7 +238,7 @@ func (r *Reconciler) runReconcileSeedFlow(
 		_ = g.Add(flow.Task{
 			Name: "Updating etcd VPA target references",
 			Fn: func(ctx context.Context) error {
-				return gardenerutils.UpdateAllEtcdVPATargetRefs(ctx, r.SeedClientSet.Client())
+				return updateAllEtcdVPATargetRefs(ctx, r.SeedClientSet.Client())
 			},
 			Dependencies: flow.NewTaskIDs(syncPointCRDs),
 		})
